sistemas_expertos/requirements.txt
Django==1.11.6
django-cors-headers==3.0.2
djangorestframework==3.9.4
pytz==2019.1
